refactor(api): split minimal api setup by usage group
